+++ src/marathon_tree.h
@@ -19,26 +19,26 @@ void marathon_tree_cleanup();
 
 // Create a new user and add him as child of parent.
 // Takes constant time.
-bool marathon_tree_add(int parentID, int userID);
+bool marathon_tree_add(unsigned int parentID, unsigned int userID);
 
 // Remove the user from the tree.
 // Takes constant time.
-bool marathon_tree_remove(int userID);
+bool marathon_tree_remove(unsigned int userID);
 
 // Add the given movie to the user's movie_list
 // Time proportional to the number of preferences of the user.
-bool marathon_tree_add_movie(int userID, int movieRating);
+bool marathon_tree_add_movie(unsigned int userID, long movieRating);
 
 // Remove the given movie from the user's movie_list
 // Time proportional to the number of preferences of the user.
-bool marathon_tree_remove_movie(int userID, int movieRating);
+bool marathon_tree_remove_movie(unsigned int userID, long movieRating);
 
 // Gives a list of at most k movies that are chosen from:
 // - All the user's preferences
 // - Results of the marathon function for its children, but only movies that
 //   have higher ratings than all of the original user's ratings are considered.
 // Time proportional to k * size of the tree.
-dlist_t *marathon_tree_get_marathon_list(int userID, int k);
+dlist_t *marathon_tree_get_marathon_list(unsigned int userID, long k);
 
 
 #endif //IPP_MARATHON_MARATHON_TREE_H
